H do not render banner when not needed
