+++ src/components/layout/MainMenu/index.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { useLocation } from 'react-router';
 
 import MainMenu from './MainMenu';
@@ -6,7 +6,19 @@ import Banner from './Banner';
 import ManageMenu from './ManageMenu';
 
 const Header = () => {
+  const [gotTxLimit, setGotTxLimit] = useState(true);
+
   const { pathname } = useLocation();
+
+  const setNotify = () => {
+    setGotTxLimit(true);
+    localStorage.setItem('gotTxLimit', 'true');
+  };
+
+  useEffect(() => {
+    const NotifyState = localStorage.getItem('gotTxLimit') === 'true';
+    setGotTxLimit(NotifyState);
+  }, []);
 
   const isVerification = pathname.startsWith('/verification');
 
@@ -21,7 +33,7 @@ const Header = () => {
     <div>
       {!isVerification && (
         <div id="header">
-          <Banner />
+          {!gotTxLimit && <Banner setNotify={setNotify} />}
           <MainMenu />
           {profileArray.includes(pathname) && <ManageMenu />}
         </div>
